Crate partiql_value

Macros

bag

list

tuple

Structs

Bag

Represents a PartiQL BAG value, e.g.: <<1, ‘two’, 4>>

BagIntoIterator

BagIter

EqualityValue

A wrapper on [T] that specifies if missing and null values should be equal.

List

Represents a PartiQL List value, e.g. [1, 2, ‘one’]

ListIntoIterator

ListIter

NullSortedValue

A wrapper on [T] that specifies if a null or missing value should be ordered before ([NULLS_FIRST] is true) or after ([NULLS_FIRST] is false) other values.

PairsIntoIter

PairsIter

Tuple

Enums

BindingIntoIter

BindingIter

BindingsName

DateTime

Value

ValueIntoIterator

ValueIter

Traits

BinaryAnd

BinaryOr

Comparable

NullableEq

NullableOrd

UnaryPlus