Enum partiql_value::Value

pub enum Value {
    Null,
    Missing,
    Boolean(bool),
    Integer(i64),
    Real(OrderedFloat<f64>),
    Decimal(Box<RustDecimal>),
    String(Box<String>),
    Blob(Box<Vec<u8>>),
    DateTime(Box<DateTime>),
    List(Box<List>),
    Bag(Box<Bag>),
    Tuple(Box<Tuple>),
}

Variants

Null

Missing

Boolean(bool)

Integer(i64)

Real(OrderedFloat<f64>)

Decimal(Box<RustDecimal>)

String(Box<String>)

Blob(Box<Vec<u8>>)

DateTime(Box<DateTime>)

List(Box<List>)

Bag(Box<Bag>)

Tuple(Box<Tuple>)

Implementations

impl Value

pub fn is_tuple(&self) -> bool

pub fn is_list(&self) -> bool

pub fn is_bag(&self) -> bool

pub fn is_sequence(&self) -> bool

pub fn is_number(&self) -> bool

Returns true if and only if Value is an integer, real, or decimal

pub fn is_absent(&self) -> bool

Returns true if and only if Value is null or missing

pub fn is_present(&self) -> bool

Returns true if Value is neither null nor missing

pub fn is_ordered(&self) -> bool

pub fn coerce_into_tuple(self) -> Tuple

pub fn coerce_to_tuple(&self) -> Tuple

pub fn as_tuple_ref(&self) -> Cow<'_, Tuple>

pub fn as_bindings(&self) -> BindingIter<'_>

pub fn into_bindings(self) -> BindingIntoIter

pub fn coerce_into_bag(self) -> Bag

pub fn as_bag_ref(&self) -> Cow<'_, Bag>

pub fn coerce_into_list(self) -> List

pub fn as_list_ref(&self) -> Cow<'_, List>

pub fn iter(&self) -> ValueIter<'_>

pub fn sequence_iter(&self) -> Option<ValueIter<'_>>